edit chamfer matching method

+++ tool_model/include/tool_model_lib/tool_model.h
@@ -167,7 +167,9 @@ public:
     cv::Point2d reproject(const cv::Mat &point, const cv::Mat &P);
 
     double calculateMatchingScore(cv::Mat &toolImage, const cv::Mat &segmentedImage, cv::Rect &ROI);
-    double calculateChamferSocre(cv::Mat &toolImage, const cv::Mat &segmentedImage, cv::Rect &ROI);
+    float calculateChamferSocre(cv::Mat &toolImage, const cv::Mat &segmentedImage, cv::Rect &ROI);
+
+
 
     /**************compute silhouettes*****************/
